pyproject.toml
[project]
name = "lidar-sidewalks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "streamlit",
    "requests",
    "geopandas",
    "shapely",
    "python-dotenv",
    "jupyterlab",
    "notebook",
    "ipython",
    "plotly",
    "ckanapi>=4.8",
    "pyarrow>=20.0.0",
    "open3d",
    "laspy[lazrs,laszip]",
    "pyvista[jupyter]",
    "trame",
    "universal_pathlib",
    "webdav4[fsspec]",
    "pdal",
]
